+++ Changelog.md
@@ -9,6 +9,7 @@
 - Add commands `/ub select[li][p] <item name>` for selecting items
 - Add expressions getcharattribute_base[int] and getcharattribute_buffed[int]
 - Fix autotinker bug where skill lowered hangs up autotinker and requires relog.  Stop button now fixes.
+- Add expressions: getdatetimelocal[], getdatetimeutc[]
 
 ## 0.2.4 (2021-03-29) [UtilityBeltInstaller-0.2.4.exe](https://gitlab.com/utilitybelt/utilitybelt.gitlab.io/uploads/deac20cf3c483c039174e4e9fee2f7ca/UtilityBeltInstaller-0.2.4.exe)
 - Fix classic server vtank patch, no longer patches non-itemenchantment auras